requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Cython==0.29.23
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
gensim==3.6.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.0.1
numpy==1.22.2
requests==2.27.1
scipy==1.8.0
smart-open==5.2.1
soupsieve==2.3.1
SQLAlchemy==1.4.31
summarization==0.0.1
urllib3==1.26.8
Werkzeug==2.0.3
